# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoBiz business operating system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business:
    get:
      operationId: getBusiness
      tags: [dashboard]
      summary: Get current business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [dashboard]
      summary: Update current business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: category
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive] }
        - name: page
          in: query
          schema: { type: number, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: number, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Product page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Deactivate a product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Product deactivated
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List product categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: page
          in: query
          schema: { type: number, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: number, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Customer page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPage"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List recent sales
      parameters:
        - name: range
          in: query
          schema: { type: string, enum: [today, week, month], default: month }
      responses:
        "200":
          description: Recent sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale and update inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report by day
      parameters:
        - name: range
          in: query
          schema: { type: string, enum: [week, month, quarter], default: month }
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportPoint"
  /assistant/chat:
    post:
      operationId: sendAssistantMessage
      tags: [assistant]
      summary: Ask AutoBiz AI about the business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    Search:
      name: search
      in: query
      schema: { type: string }
  schemas:
...
    Business:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        city: { type: string }
        currency: { type: string }
      required: [id, name, type, city, currency]
    BusinessUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string, minLength: 1 }
        city: { type: string, minLength: 1 }
        currency: { type: string, minLength: 3, maxLength: 3 }
      required: [name, type, city, currency]
...
    Category:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        productCount: { type: number }
        color: { type: string }
      required: [id, name, productCount, color]
    CategoryInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        color: { type: string }
      required: [name]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        brand: { type: string }
        unit: { type: string }
        purchasePrice: { type: number }
        sellingPrice: { type: number }
        currentStock: { type: number }
        minimumStock: { type: number }
        status: { type: string, enum: [active, inactive] }
        updatedAt: { type: string }
      required: [id, name, sku, category, brand, unit, purchasePrice, sellingPrice, currentStock, minimumStock, status, updatedAt]
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string, minLength: 1 }
        category: { type: string }
        brand: { type: string }
        unit: { type: string }
        purchasePrice: { type: number, minimum: 0 }
        sellingPrice: { type: number, minimum: 0 }
        currentStock: { type: number, minimum: 0 }
        minimumStock: { type: number, minimum: 0 }
      required: [name, sku, category, brand, unit, purchasePrice, sellingPrice, currentStock, minimumStock]
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        brand: { type: string }
        unit: { type: string }
        purchasePrice: { type: number, minimum: 0 }
        sellingPrice: { type: number, minimum: 0 }
        currentStock: { type: number, minimum: 0 }
        minimumStock: { type: number, minimum: 0 }
        status: { type: string, enum: [active, inactive] }
    ProductPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        total: { type: number }
        page: { type: number }
        pageSize: { type: number }
      required: [items, total, page, pageSize]
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        city: { type: string }
        totalPurchases: { type: number }
        outstanding: { type: number }
        lastPurchaseAt: { type: ["string", "null"] }
      required: [id, name, phone, email, city, totalPurchases, outstanding, lastPurchaseAt]
    CustomerInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        email: { type: string }
        city: { type: string }
      required: [name, phone, email, city]
    CustomerUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        email: { type: string }
        city: { type: string }
    CustomerPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Customer" }
        total: { type: number }
        page: { type: number }
        pageSize: { type: number }
      required: [items, total, page, pageSize]
    Supplier:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        city: { type: string }
        payable: { type: number }
        lastOrderAt: { type: ["string", "null"] }
      required: [id, name, phone, city, payable, lastOrderAt]
    Sale:
      type: object
      properties:
        id: { type: string }
        invoiceNumber: { type: string }
        customerName: { type: string }
        total: { type: number }
        paymentStatus: { type: string, enum: [paid, partial, pending] }
        itemCount: { type: number }
        createdAt: { type: string }
      required: [id, invoiceNumber, customerName, total, paymentStatus, itemCount, createdAt]
    SaleItemInput:
      type: object
      properties:
        productId: { type: string }
        quantity: { type: number, minimum: 1 }
        unitPrice: { type: number, minimum: 0 }
      required: [productId, quantity, unitPrice]
    SaleInput:
      type: object
      properties:
        customerId: { type: ["string", "null"] }
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/SaleItemInput" }
        paymentStatus: { type: string, enum: [paid, partial, pending] }
      required: [customerId, items, paymentStatus]
...
    ReportPoint:
      type: object
      properties:
        label: { type: string }
        value: { type: number }
      required: [label, value]
    AssistantMessageInput:
      type: object
      properties:
        message: { type: string, minLength: 1 }
        conversationId: { type: ["string", "null"] }
      required: [message, conversationId]
    AssistantResponse:
      type: object
      properties:
        message: { type: string }
        conversationId: { type: string }
        intent: { type: string }
        facts:
          type: array
          items: { type: string }
      required: [message, conversationId, intent, facts]